New files: read_gc_nd51.py and cleanup_py_nd51_struct. Updates to geos_chem_delAMF
NEW:
    read_gc_nd51.py - Python script to read the ND51 bpch files
output from GEOS-Chem.  This was easier to do in Python to loop
over all days because of the bpch module that can treat those files
directly. Note that the bpch module is not installed in default
Python installations and so needs to be added.

    cleanup_up_nd51_struct.m - Matlab script to clean up the output
from read_gc_nd51.py a little bit. Both reorders the structure fields
to be in the order we expect (i.e. the same as read_geos_output.m
gives us) and allows concatenation of multiple structure together.

MODIFIED:
   geos_chem_delAMF.m - now includes an optional fifth argument,
which is a logical matrix that has the same first two dimensions
as the input profile matrices.  This allows the user to skip
certain pixels in order to speed up processing and focus only
on those pixels of interest.